+++ src/main/java/pl/edu/pwr/pwrinspace/poliwrocket/Controller/ValvesPressurizingController.java
@@ -1,0 +1,4 @@
+package pl.edu.pwr.pwrinspace.poliwrocket.Controller;
+
+public class ValvesPressurizingController extends BaseButtonSensorCommandsController {
+}
