--- src/main/java/pl/edu/pwr/pwrinspace/poliwrocket/Controller/ValvesTimeOpenController.java
@@ -1,4 +1,0 @@
-package pl.edu.pwr.pwrinspace.poliwrocket.Controller;
-
-public class ValvesTimeOpenController extends BaseButtonSensorCommandsController {
-}
