--- src/main/java/pl/edu/pwr/pwrinspace/poliwrocket/Controller/ValvesTimeOpenPressurizingController.java
@@ -1,4 +1,0 @@
-package pl.edu.pwr.pwrinspace.poliwrocket.Controller;
-
-public class ValvesTimeOpenPressurizingController extends BaseButtonSensorCommandsController {
-}
